+++ lib/sus/fixtures/protocol/http.rb
@@ -5,4 +5,4 @@
 
 require_relative "http/version"
 require_relative "http/client"
-require_relative "http/application_context"
+require_relative "http/middleware_context"