+++ releases.md
@@ -2,4 +2,4 @@
 
 ## Unreleased
 
-  - Introduce an in-process client and application context for protocol HTTP middleware.
+  - Introduce an in-process client and middleware context for protocol HTTP middleware.